Uses a new name for the OFBiz logo file easy to reuse everywhere and in future.
There are other places where the  OFBiz logo file is used, at least in a lot (if not all) README.md files. If we change the  OFBiz logo file name at every  new versions we break these links. Previous name,logo.png, was poorly chosen because it's ambiguous. Current name OFBiz-logoV3-apache.png introduces a version in the name which is not reusable in future. By using OFBiz-logo.png we define a unambiguous name, easily reusable everywhere and in future. 

Conventions make things easier, here the OFBiz logo file name is a convention.

git-svn-id: http://svn.apache.org/repos/asf/ofbiz/site@1813479 13f79535-47bb-0310-9956-ffa450edef68