Improved: Add subscribe email on website mailing list page (OFBIZ-10729)
Currently, we have a link of lists.apache.org for subscribing mailing list,
When a user visited lists.apache.org website it has the option to subscribe, this tries to open email client, if no email client setup on a user machine, so its difficult to subscribe.

We can add an alternate option as well along with lists.apache.org link. A simple text with subscribe email address.