Improved: definitely removes most social networks in header
They were only commented out for now.
This also completes the removing of Tweeter from News menu (not only template)

Reminder:
This is related to this thread:
https://lists.apache.org/thread/zwzs0bqfm9cb6rqnl05q0hyphc731odx

Are banned Twitter(now X), Linkedin and Facebook for reasons explained in the thread

So only remains Youtube that is really helpful for user

The show off time is done. Rather each service providers should do it.