improved: Have an XSD regarding portal records (OFBIZ-13097)
Per suggestion by Jacques Le Roux:
But we really miss the same for data used to define portals. So I strongly suggest that we create a shema for *PortletData.xml files. It should not be hard to create. It would help possible customers to create their own portals with a bit of documentation that seriously lacks for now.

added: ofbiz-portal.xsd