Implemented: trying a last thing before creating an issue at https://github.com/apache/infrastructure-asfyaml/issues
As can be seen at https://github.com/apache/ofbiz-site/commits/master/,
since Mar 25, 2025 I have been unable to update the OFBiz
[widget-form.xsd](https://ofbiz.apache.org/dtds/) on the
[OFBiz site](https://ofbiz.apache.org/). The last update is the 2025-02-23.

Could it be because  [.asf.yaml is not changed on site]
(https://ofbiz.apache.org/.asf.yaml).
And if so why is this file not updated? Should I delete and recreate it.

OK, trying that...